% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-quasirandom.R
\name{geom_quasirandom}
\alias{geom_quasirandom}
\title{Quasirandom points}
\usage{
geom_quasirandom(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  ...,
  method = "quasirandom",
  width = NULL,
  varwidth = FALSE,
  bandwidth = 0.5,
  nbins = NULL,
  dodge.width = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{method}{\code{string}. Method for arranging points, default is
\code{"quasirandom"}. See details below.}

\item{width}{\code{numeric}. Maximum spacing away from the centre for each group
of points, default is \code{1}.}

\item{varwidth}{\code{boolean}. Adjust the width of each group based on the number
of points in the group. Default is \code{FALSE}.}

\item{bandwidth}{\code{numeric}. Specifies the bandwidth used to calculate the
kernel density, default is \code{1}. Smaller values = tighter fit. larger values
= looser fit.}

\item{nbins}{\code{integer} The number of points used to calculate density
(default is \code{1000} for quasirandom and pseudorandom, and 100 for others).}

\item{dodge.width}{\code{numeric}. Amount to dodge points from different aesthetic
groups, default is \code{NULL} for no dodging.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
The quasirandom geom is a convenient shortcut for
\code{geom_point(position = "quasirandom")}. It shifts points to avoid
overplotting using the \code{\link{offsetSingleGroup}} function from the
\code{vipor} package.
}
\details{
\strong{method:} specifies the algorithm used to distribute the points.

\code{"quasirandom"}: points are distributed within a kernel density estimate of
the distribution with offset determined by quasirandom Van de Corput noise.

\code{"pseudorandom"}: points are distributed within a kernel density estimate of
the distribution with offset determined by pseudorandom noise a la jitter.

\code{"maxout"}: points are distributed within a kernel density with points in a
band distributed with highest value points on the outside and lowest in
the middle.

\code{"minout"}: points are distributed within a kernel density with points in a
band distributed with highest value points in the middle and lowest on
the outside.

\code{"tukey"}: points are distributed as described in Tukey and Tukey,
"Strips displaying empirical distributions: I. textured dot strips".

\code{"tukeyDense"}: points are distributed as described in Tukey and Tukey but
are constrained with the kernel density estimate.
}
\section{Aesthetics}{

\code{geom_quasirandom()} understands the following aesthetics (required aesthetics
are in bold):
\itemize{
\item \strong{x}
\item \strong{y}
\item alpha
\item colour
\item fill
\item group
\item shape
\item size
\item stroke
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\examples{
#
}
\seealso{
\code{\link[=position_quasirandom]{position_quasirandom()}} for the underlying function to this geom.
\code{\link[=geom_beeswarm]{geom_beeswarm()}} for another method of shifting points to avoid
overplotting.
}
