% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbeeswarm-package.R
\docType{package}
\name{ggbeeswarm2-package}
\alias{ggbeeswarm2}
\alias{ggbeeswarm2-package}
\title{ggbeeswarm2: Categorical Scatter (Violin Point) Plots}
\description{
Provides two methods of plotting categorical scatter plots such
    that the arrangement of points within a category reflects the density of
    data at that region, and avoids over-plotting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csdaw/ggbeeswarm2}
  \item Report bugs at \url{https://github.com/csdaw/ggbeeswarm2/issues}
}

}
\author{
\strong{Maintainer}: Charlotte Dawson \email{csdaw@outlook.com}

Authors:
\itemize{
  \item Erik Clarke \email{erikclarke@gmail.com}
  \item Scott Sherrill-Mix \email{sherrillmix@gmail.com}
}

}
\keyword{internal}
