% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-quasirandom.R
\name{position_quasirandom}
\alias{position_quasirandom}
\title{Separate coincident points with the vipor package}
\usage{
position_quasirandom(
  method = "quasirandom",
  width = NULL,
  varwidth = FALSE,
  bandwidth = 0.5,
  nbins = NULL,
  dodge.width = NULL
)
}
\arguments{
\item{method}{\code{string}. Method for arranging points, default is
\code{"quasirandom"}. See details below.}

\item{width}{\code{numeric}. Maximum spacing away from the centre for each group
of points, default is \code{1}.}

\item{varwidth}{\code{boolean}. Adjust the width of each group based on the number
of points in the group. Default is \code{FALSE}.}

\item{bandwidth}{\code{numeric}. Specifies the bandwidth used to calculate the
kernel density, default is \code{1}. Smaller values = tighter fit. larger values
= looser fit.}

\item{nbins}{\code{integer} The number of points used to calculate density
(default is \code{1000} for quasirandom and pseudorandom, and 100 for others).}

\item{dodge.width}{\code{numeric}. Amount to dodge points from different aesthetic
groups, default is \code{NULL} for no dodging.}
}
\description{
Offset points to avoid overplotting using the
\code{\link{offsetSingleGroup}} function from the \code{vipor} package.
}
\details{
\strong{method:} specifies the algorithm used to distribute the points.

\code{"quasirandom"}: points are distributed within a kernel density estimate of
the distribution with offset determined by quasirandom Van de Corput noise.

\code{"pseudorandom"}: points are distributed within a kernel density estimate of
the distribution with offset determined by pseudorandom noise a la jitter.

\code{"maxout"}: points are distributed within a kernel density with points in a
band distributed with highest value points on the outside and lowest in
the middle.

\code{"minout"}: points are distributed within a kernel density with points in a
band distributed with highest value points in the middle and lowest on
the outside.

\code{"tukey"}: points are distributed as described in Tukey and Tukey,
"Strips displaying empirical distributions: I. textured dot strips".

\code{"tukeyDense"}: points are distributed as described in Tukey and Tukey but
are constrained with the kernel density estimate.
}
\examples{
#
}
\seealso{
\code{\link[=position_beeswarm]{position_beeswarm()}}
}
